chore: rename error subclasses directory
