+++ src/core/errors/classes/local-config-not-found-solo-error.ts
