+++ src/core/errors/classes/remote-configs-mismatch-solo-error.ts
